+++ packages/core/src/actions/send.ts
@@ -1,28 +1,124 @@
 import isDevelopment from '#is-development';
+import { constantPrefixes, error } from '../actions.ts';
 import {
+  ActionArgs,
+  ActorRef,
+  AnyActorContext,
+  AnyActorRef,
+  AnyEventObject,
+  AnyInterpreter,
+  AnyState,
+  Cast,
+  DelayExpr,
+  EventFrom,
   EventObject,
+  InferEvent,
+  MachineContext,
+  SendExpr,
+  SendToActionOptions,
   SendToActionParams,
   SpecialTargets,
-  SendExpr,
-  AnyEventObject,
-  MachineContext
-} from '../types.ts';
-import { sendTo as sendToActionType } from '../actionTypes.ts';
-import { isFunction, isString } from '../utils.ts';
-import { createDynamicAction } from '../../actions/dynamicAction.ts';
-import {
-  ActorRef,
-  AnyActorRef,
-  AnyInterpreter,
-  BaseDynamicActionObject,
-  Cast,
-  EventFrom,
-  InferEvent,
-  SendToActionObject,
-  SendToActionOptions,
   UnifiedArg
-} from '../index.ts';
-import { actionTypes, error } from '../actions.ts';
+} from '../types.ts';
+
+function resolve(
+  actorContext: AnyActorContext,
+  state: AnyState,
+  args: ActionArgs<any, any>,
+  {
+    to,
+    event: eventOrExpr,
+    id,
+    delay
+  }: {
+    to:
+      | AnyActorRef
+      | string
+      | ((
+          args: UnifiedArg<MachineContext, EventObject>
+        ) => AnyActorRef | string);
+    event: EventObject | SendExpr<MachineContext, EventObject, EventObject>;
+    id: string | undefined;
+    delay: string | number | DelayExpr<MachineContext, EventObject> | undefined;
+  }
+) {
+  const delaysMap = state.machine.implementations.delays;
+
+  if (typeof eventOrExpr === 'string') {
+    throw new Error(
+      `Only event objects may be used with sendTo; use sendTo({ type: "${eventOrExpr}" }) instead`
+    );
+  }
+  const resolvedEvent =
+    typeof eventOrExpr === 'function' ? eventOrExpr(args) : eventOrExpr;
+
+  let resolvedDelay: number | undefined;
+  if (typeof delay === 'string') {
+    const configDelay = delaysMap && delaysMap[delay];
+    resolvedDelay =
+      typeof configDelay === 'function' ? configDelay(args) : configDelay;
+  } else {
+    resolvedDelay = typeof delay === 'function' ? delay(args) : delay;
+  }
+
+  const resolvedTarget = typeof to === 'function' ? to(args) : to;
+  let targetActorRef: AnyActorRef | undefined;
+
+  if (typeof resolvedTarget === 'string') {
+    if (resolvedTarget === SpecialTargets.Parent) {
+      targetActorRef = actorContext?.self._parent;
+    } else if (resolvedTarget === SpecialTargets.Internal) {
+      targetActorRef = actorContext?.self;
+    } else if (resolvedTarget.startsWith('#_')) {
+      // SCXML compatibility: https://www.w3.org/TR/scxml/#SCXMLEventProcessor
+      // #_invokeid. If the target is the special term '#_invokeid', where invokeid is the invokeid of an SCXML session that the sending session has created by <invoke>, the Processor must add the event to the external queue of that session.
+      targetActorRef = state.children[resolvedTarget.slice(2)];
+    } else {
+      targetActorRef = state.children[resolvedTarget];
+    }
+    if (!targetActorRef) {
+      throw new Error(
+        `Unable to send event to actor '${resolvedTarget}' from machine '${state.machine.id}'.`
+      );
+    }
+  } else {
+    targetActorRef = resolvedTarget || actorContext?.self;
+  }
+
+  return [
+    state,
+    { to: targetActorRef, event: resolvedEvent, id, delay: resolvedDelay }
+  ];
+}
+function execute(
+  actorContext: AnyActorContext,
+  params: {
+    to: AnyActorRef;
+    event: EventObject;
+    id: string | undefined;
+    delay: number | undefined;
+  }
+) {
+  if (typeof params.delay === 'number') {
+    (actorContext.self as AnyInterpreter).delaySend(
+      params as typeof params & { delay: number }
+    );
+    return;
+  }
+
+  const { to, event } = params;
+
+  actorContext.defer(() => {
+    to.send(
+      event.type === constantPrefixes.error
+        ? {
+            type: `${error(actorContext.self.id)}`,
+            data: (event as any).data
+          }
+        : event
+    );
+  });
+}
 
 /**
  * Sends an event to an actor.
@@ -35,153 +131,39 @@ import { actionTypes, error } from '../actions.ts';
  */
 export function sendTo<
   TContext extends MachineContext,
+  TExpressionEvent extends EventObject,
   TEvent extends EventObject,
   TActor extends AnyActorRef
 >(
-  actor:
+  to:
     | TActor
     | string
-    | ((args: UnifiedArg<TContext, TEvent>) => TActor | string),
+    | ((args: UnifiedArg<TContext, TExpressionEvent>) => TActor | string),
   eventOrExpr:
     | EventFrom<TActor>
     | SendExpr<
         TContext,
-        TEvent,
+        TExpressionEvent,
         InferEvent<Cast<EventFrom<TActor>, EventObject>>
       >,
-  options?: SendToActionOptions<TContext, TEvent>
-): BaseDynamicActionObject<
-  TContext,
-  TEvent,
-  TEvent,
-  SendToActionObject<AnyEventObject>,
-  SendToActionParams<TContext, TEvent>
-> {
-  return createDynamicAction<
-    TContext,
-    TEvent,
-    TEvent,
-    SendToActionObject<AnyEventObject>,
-    SendToActionParams<TContext, TEvent>
-  >(
-    {
-      type: sendToActionType,
-      params: {
-        to: actor,
-        delay: options ? options.delay : undefined,
-        event: eventOrExpr,
-        id:
-          options && options.id !== undefined
-            ? options.id
-            : isFunction(eventOrExpr)
-            ? eventOrExpr.name
-            : (eventOrExpr as TEvent).type
-      }
-    },
-    (event, { actorContext, state }) => {
-      const params = {
-        to: actor,
-        delay: options ? options.delay : undefined,
-        event: eventOrExpr,
-        // TODO: don't auto-generate IDs here like that
-        // there is too big chance of the ID collision
-        id:
-          options && options.id !== undefined
-            ? options.id
-            : isFunction(eventOrExpr)
-            ? eventOrExpr.name
-            : (eventOrExpr as TEvent).type
-      };
-      const args: UnifiedArg<TContext, TEvent> = {
-        context: state.context,
-        event,
-        self: actorContext?.self ?? (null as any),
-        system: actorContext?.system
-      };
-      const delaysMap = state.machine.implementations.delays;
-
-      // TODO: helper function for resolving Expr
-      if (typeof eventOrExpr === 'string') {
-        throw new Error(
-          `Only event objects may be used with sendTo; use sendTo({ type: "${eventOrExpr}" }) instead`
-        );
-      }
-      const resolvedEvent = isFunction(eventOrExpr)
-        ? eventOrExpr(args)
-        : eventOrExpr;
-
-      let resolvedDelay: number | undefined;
-      if (isString(params.delay)) {
-        const configDelay = delaysMap && delaysMap[params.delay];
-        resolvedDelay = isFunction(configDelay)
-          ? configDelay(args)
-          : configDelay;
-      } else {
-        resolvedDelay = isFunction(params.delay)
-          ? params.delay(args)
-          : params.delay;
-      }
-
-      const resolvedTarget = isFunction(params.to)
-        ? params.to(args)
-        : params.to;
-      let targetActorRef: AnyActorRef | undefined;
-
-      if (typeof resolvedTarget === 'string') {
-        if (resolvedTarget === SpecialTargets.Parent) {
-          targetActorRef = actorContext?.self._parent;
-        } else if (resolvedTarget === SpecialTargets.Internal) {
-          targetActorRef = actorContext?.self;
-        } else if (resolvedTarget.startsWith('#_')) {
-          // SCXML compatibility: https://www.w3.org/TR/scxml/#SCXMLEventProcessor
-          // #_invokeid. If the target is the special term '#_invokeid', where invokeid is the invokeid of an SCXML session that the sending session has created by <invoke>, the Processor must add the event to the external queue of that session.
-          targetActorRef = state.children[resolvedTarget.slice(2)];
-        } else {
-          targetActorRef = state.children[resolvedTarget];
-        }
-        if (!targetActorRef) {
-          throw new Error(
-            `Unable to send event to actor '${resolvedTarget}' from machine '${state.machine.id}'.`
-          );
-        }
-      } else {
-        targetActorRef = resolvedTarget || actorContext?.self;
-      }
-
-      const resolvedAction: SendToActionObject = {
-        type: actionTypes.sendTo,
-        params: {
-          ...params,
-          to: targetActorRef,
-          event: resolvedEvent,
-          delay: resolvedDelay
-        },
-        execute: (actorCtx) => {
-          const sendAction = resolvedAction as SendToActionObject;
-
-          if (typeof sendAction.params.delay === 'number') {
-            (actorCtx.self as AnyInterpreter).delaySend(sendAction);
-            return;
-          } else {
-            const target = sendAction.params.to!;
-            const sentEvent = sendAction.params.event;
-            actorCtx.defer(() => {
-              target.send(
-                sentEvent.type === actionTypes.error
-                  ? {
-                      type: `${error(actorCtx.self.id)}`,
-                      data: sentEvent.data
-                    }
-                  : sendAction.params.event
-              );
-            });
-          }
-        }
-      };
-
-      return [state, resolvedAction];
+  options?: SendToActionOptions<TContext, TExpressionEvent>
+) {
+  function sendTo(_: ActionArgs<TContext, TExpressionEvent>) {
+    if (isDevelopment) {
+      throw new Error(`This isn't supposed to be called`);
     }
-  );
+  }
+
+  sendTo.type = 'xstate.sendTo';
+  sendTo.to = to;
+  sendTo.event = eventOrExpr;
+  sendTo.id = options?.id;
+  sendTo.delay = options?.delay;
+
+  sendTo.resolve = resolve;
+  sendTo.execute = execute;
+
+  return sendTo;
 }
 
 /**
@@ -192,13 +174,14 @@ export function sendTo<
  */
 export function sendParent<
   TContext extends MachineContext,
+  TExpressionEvent extends EventObject,
   TEvent extends EventObject,
   TSentEvent extends EventObject = AnyEventObject
 >(
-  event: TSentEvent | SendExpr<TContext, TEvent, TSentEvent>,
-  options?: SendToActionOptions<TContext, TEvent>
+  event: TSentEvent | SendExpr<TContext, TExpressionEvent, TSentEvent>,
+  options?: SendToActionOptions<TContext, TExpressionEvent>
 ) {
-  return sendTo<TContext, TEvent, AnyActorRef>(
+  return sendTo<TContext, TExpressionEvent, TEvent, AnyActorRef>(
     SpecialTargets.Parent,
     event,
     options
@@ -218,10 +201,10 @@ type Target<TContext extends MachineContext, TEvent extends EventObject> =
  */
 export function forwardTo<
   TContext extends MachineContext,
-  TEvent extends EventObject
+  TExpressionEvent extends EventObject
 >(
-  target: Target<TContext, TEvent>,
-  options?: SendToActionOptions<TContext, TEvent>
+  target: Target<TContext, TExpressionEvent>,
+  options?: SendToActionOptions<TContext, TExpressionEvent>
 ) {
   if (isDevelopment && (!target || typeof target === 'function')) {
     const originalTarget = target;
@@ -238,7 +221,7 @@ export function forwardTo<
       return resolvedTarget;
     };
   }
-  return sendTo<TContext, TEvent, AnyActorRef>(
+  return sendTo<TContext, TExpressionEvent, EventObject, AnyActorRef>(
     target,
     ({ event }: any) => event,
     options
@@ -254,16 +237,19 @@ export function forwardTo<
  */
 export function escalate<
   TContext extends MachineContext,
-  TEvent extends EventObject,
+  TExpressionEvent extends EventObject,
   TErrorData = any
 >(
-  errorData: TErrorData | ((args: UnifiedArg<TContext, TEvent>) => TErrorData),
-  options?: SendToActionParams<TContext, TEvent>
+  errorData:
+    | TErrorData
+    | ((args: UnifiedArg<TContext, TExpressionEvent>) => TErrorData),
+  options?: SendToActionParams<TContext, TExpressionEvent>
 ) {
-  return sendParent<TContext, TEvent>((arg) => {
+  return sendParent<TContext, TExpressionEvent, EventObject>((arg) => {
     return {
-      type: actionTypes.error,
-      data: isFunction(errorData) ? errorData(arg) : errorData
+      type: constantPrefixes.error,
+      data:
+        typeof errorData === 'function' ? (errorData as any)(arg) : errorData
     };
   }, options);
 }