+++ packages/core/test/interpreter.test.ts
@@ -25,11 +25,11 @@ const lightMachine = createMachine({
   initial: 'green',
   states: {
     green: {
-      entry: [raise({ type: 'TIMER' }, { delay: 10 })],
+      entry: [raise({ type: 'TIMER' }, { id: 'TIMER1', delay: 10 })],
       on: {
         TIMER: 'yellow',
         KEEP_GOING: {
-          actions: [cancel('TIMER')]
+          actions: [cancel('TIMER1')]
         }
       }
     },