+++ packages/xstate-immer/src/index.ts
@@ -1,24 +1,15 @@
+import { Draft, produce } from 'immer';
 import {
-  EventObject,
-  BaseActionObject,
-  assign as xstateAssign,
   AssignArgs,
+  EventObject,
   MachineContext,
-  DynamicAssignAction
+  assign as xstateAssign
 } from 'xstate';
-import { produce, Draft } from 'immer';
 
 export type ImmerAssigner<
   TContext extends MachineContext,
   TExpressionEvent extends EventObject
 > = (args: AssignArgs<Draft<TContext>, TExpressionEvent>) => void;
-
-export interface ImmerAssignAction<
-  TContext extends MachineContext,
-  TExpressionEvent extends EventObject
-> extends BaseActionObject {
-  assignment: ImmerAssigner<TContext, TExpressionEvent>;
-}
 
 function immerAssign<
   TContext extends MachineContext,
@@ -49,22 +40,10 @@ export interface ImmerUpdateEvent<
   input: TInput;
 }
 
-export interface ImmerUpdater<
-  TContext extends MachineContext,
-  TEvent extends ImmerUpdateEvent
-> {
-  update: (input: TEvent['input']) => TEvent;
-  action: DynamicAssignAction<TContext, TEvent>;
-  type: TEvent['type'];
-}
-
 export function createUpdater<
   TContext extends MachineContext,
   TEvent extends ImmerUpdateEvent
->(
-  type: TEvent['type'],
-  recipe: ImmerAssigner<TContext, TEvent>
-): ImmerUpdater<TContext, TEvent> {
+>(type: TEvent['type'], recipe: ImmerAssigner<TContext, TEvent>) {
   const update = (input: TEvent['input']): TEvent => {
     return {
       type,